+++ src/SignalR.Orleans/Constants.cs
@@ -19,5 +19,6 @@ namespace SignalR.Orleans
         internal const int STREAM_SEND_REPLICAS = 10; // todo: make configurable instead
         internal const double HEARTBEAT_PULSE_IN_MINUTES = 30;
         internal const double SERVERDIRECTORY_CLEANUP_IN_MINUTES = HEARTBEAT_PULSE_IN_MINUTES * 3;
+        internal const string CONNECTION_STREAM_CLEANUP = "0:01:00";
     }
 }